% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdriver_functions.R
\name{calculate_rate}
\alias{calculate_rate}
\title{Calculate sampling rate based on cohort size}
\usage{
calculate_rate(total_cohort, synonymous_sampled = 10, max_rate = 0.202)
}
\arguments{
\item{total_cohort}{Numeric. Total number of samples in the cohort.}

\item{synonymous_sampled}{Numeric. Desired number of synonymous mutations sampled
at 5\% of the cohort (10).}

\item{max_rate}{Numeric. Maximum rate (0.202).}
}
\value{
A numeric rate.
}
\description{
Computes a rate such that at 5\% of the total cohort size, the desired number
of synonymous mutations is achieved. The rate is chosen as the minimum of the rate
computed from the desired number of synonymous mutations and 0.202 which ensures at least
one synonymous mutation is sampled per patient when only 2 nonsynonymous mutations are present
}
