% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdriver_functions.R
\name{compute_rna_depth_weights}
\alias{compute_rna_depth_weights}
\title{Compute RNA depth weights for all mutations}
\usage{
compute_rna_depth_weights(mutation_table, cancer_type, outdir)
}
\arguments{
\item{mutation_table}{Input mutation table}

\item{cancer_type}{Character. Cancer type label used in plotting and saving}

\item{outdir}{Character. Directory in which to save the depth plot.}
}
\value{
The updated mutation table with added columns:
\code{RNA_depth_capped}, \code{log_cov}, \code{rna_vaf_weight}, and \code{rna_vaf_max_weight}.
}
\description{
Computes mutation-specific weights based on RNA depth. The function:
\enumerate{
\item Caps RNA depth at the 90th percentile of RNA depths across the cohort
\item Adds a new column (\code{RNA_depth_capped}) for the capped depth.
\item Generates and saves a histogram of the RNA depth distribution.
\item Computes \code{log_cov} as \code{log2(RNA_depth_capped + 1)}.
\item Computes weights (\code{rna_vaf_weight} and \code{rna_vaf_max_weight})
using helper functions.
\item Replaces zeros in \code{rna_vaf_max_weight} as if they had a coverage of 1
}
}
