% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdriver_functions.R
\name{create_synonymous_background_including_mut_type}
\alias{create_synonymous_background_including_mut_type}
\title{Create Synonymous Background (with mutation-type sampling)}
\usage{
create_synonymous_background_including_mut_type(
  syn_df,
  gene_df,
  synon_threshold,
  synonymous_background_filtered,
  seed,
  non_synon_muts_list,
  sampling_method,
  include_gene_synons = FALSE
)
}
\arguments{
\item{syn_df}{Data frame of synonymous mutations for relevant patients.}

\item{gene_df}{Data frame of non-synonymous mutations for the gene of interest}

\item{synon_threshold}{Integer specifying how many synonymous mutations to try to sample.}

\item{seed}{An integer seed for reproducibility.}

\item{non_synon_muts_list}{Named list of data frames of non-synonymous mutations by patient.}

\item{sampling_method}{"strict" or "relaxed" - Default: relaxed}

\item{include_gene_synons}{Logical; if \code{TRUE}, Synonymous mutations from the gene of interest are also sampled}

\item{synonymous_background}{A data frame of all synonymous mutations across the cohort}
}
\value{
A list with elements:
\itemize{
\item \code{final_synonymous_background}: the sampled synonymous mutations
\item \code{sampling_summary_df}: a data frame with summary of from where synonymous mutations were sampled
\item \code{proportion_sampled_per_patient}: fraction of required synonymous mutations that actually got sampled for each patient
}
}
\description{
\code{create_synonymous_background_including_mut_type} attempts to sample enough
synonymous mutations per tumour - in a heirarchical order:
Consider synonymous mutations only within the tumour harbouring a mutation within the gene of interest
Consider synonymous mutations within all tumours harbouring a mutation within the gene of interest
Consider all synonymous mutations from tumours within the cancer type being considered
Consider mutations of the same type from within that sample.
If sufficient mutations have not been sampled at this point, proceed with the incompletely sampled background.
}
