% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdriver_functions.R
\name{exponential_growth_limited_2x}
\alias{exponential_growth_limited_2x}
\title{Exponential growth sampling capped at 2x sample size}
\usage{
exponential_growth_limited_2x(n_samples, base_num = 1, rate)
}
\arguments{
\item{n_samples}{Numeric. Number of samples.}

\item{base_num}{Numeric. The base number for the exponential calculation (default is 1).}

\item{rate}{Numeric. The growth rate.}
}
\value{
An integer sample size.
}
\description{
Computes the number of synonymous mutations to be sampled per test.
}
