% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdriver_functions.R
\name{get_rna_weights}
\alias{get_rna_weights}
\title{Compute RNA Weights}
\usage{
get_rna_weights(coverage)
}
\arguments{
\item{coverage}{A numeric vector of RNA depth of coverage}
}
\value{
A numeric vector of weights
}
\description{
\code{get_rna_weights} computes a per-variant weight based on the fraction of
reads for the variant, against all other variants
}
