% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdriver_functions.R
\name{myTryCatch}
\alias{myTryCatch}
\title{Try-Catch Helper}
\usage{
myTryCatch(expr)
}
\arguments{
\item{expr}{Expression being evaluated}
}
\value{
A list with elements \code{value}, \code{warning}, and \code{error}.
}
\description{
\code{myTryCatch} is a helper function that captures warnings, errors, and
returns them in a list.
}
