% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdriver_functions.R
\name{perform_synonymous_scaling}
\alias{perform_synonymous_scaling}
\title{Perform synonymous mutation scaling and output plots}
\usage{
perform_synonymous_scaling(
  synonymous_background,
  mutation_table,
  non_synonymous_key,
  genes,
  cgc_list,
  synon_depth_filt,
  cancer_type,
  outdir
)
}
\arguments{
\item{synonymous_background}{A data frame containing the synonymous mutations}

\item{mutation_table}{A data frame with all mutations.}

\item{non_synonymous_key}{A character vector of non-synonymous mutation types.}

\item{genes}{A data frame of genes of interest}

\item{cgc_list}{A data frame of CGC genes.}

\item{synon_depth_filt}{Numeric. The depth threshold for synonymous mutations.}

\item{cancer_type}{Character. Cancer type label}

\item{outdir}{Character. Directory in which to save the scaling plots.}
}
\value{
A list with two elements:
\itemize{
\item \code{plot}: A combined ggplot object comparing results before and after scaling.
\item \code{synonymous_background}: The updated synonymous background table with scaled \code{RNA_VAF}.
}
}
\description{
Performs scaling of synonymous mutations to match the average RNA VAFs of missense mutations
}
