% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdriver_functions.R
\name{plot_qq}
\alias{plot_qq}
\title{Plot QQ Plot for RVdriver Results}
\usage{
plot_qq(data, method_name, outdir)
}
\arguments{
\item{data}{A data frame containing the RVdriver results}

\item{method_name}{Character. The name of the method to plot.}

\item{outdir}{Character. Directory in which to save the QQ plot.}
}
\value{
QQ plot
}
\description{
Generates a QQ plot comparing the expected and observed -log10 p-values for RVdriver results
The plot is saved to the output directory.
}
