% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdriver_functions.R
\name{rvdriver}
\alias{rvdriver}
\title{Main RVdriver Function}
\usage{
rvdriver(
  comp_df,
  seed_list,
  synon_threshold,
  num_iter,
  g,
  ncores,
  non_synon_muts_list,
  sampling_method,
  include_gene_synons,
  rounding_method
)
}
\arguments{
\item{comp_df}{Data frame including all mutations to be utilised in each iteration.}

\item{seed_list}{A vector of seeds for reproducibility across iterations.}

\item{synon_threshold}{Number of synonymous mutations to be sampled per sample}

\item{num_iter}{Integer: number of iterations.}

\item{g}{The gene name.}

\item{ncores}{Number of cores for mclapply parellelisation}

\item{non_synon_muts_list}{Named list of data frames of nonsynonymous mutations by patient.}

\item{sampling_method}{\code{"strict"} or \code{"relaxed"} approach. default: relaxed}

\item{include_gene_synons}{Logical; if \code{TRUE}, try sampling synons from same gene. default: TRUE}

\item{rounding_method}{Method for determining number of synonymous mutations to be sampled (e.g. \code{"expo"}).}
}
\value{
A list with two elements:
\itemize{
\item \code{res_df}: a data frame summarising the results
\item \code{proportion_sampled_per_patient}: a data frame summarising how many
synonyous mutations got sampled per patient
}
}
\description{
\code{rvdriver} is the main function that samples synonymous mutations per sample,
computing an lm(RNA_VAF ~ mutation_function, weights = mutation_depth),
(via \code{parallel_function}) across multiple iterations, returning summary statistics
per gene.
}
