% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdriver_functions.R
\name{sample_and_update}
\alias{sample_and_update}
\title{Sample and Update Synonymous Mutations}
\usage{
sample_and_update(patient_ids, threshold, source_df_name, source_name, missing)
}
\arguments{
\item{patient_ids}{Vector of patient IDs.}

\item{threshold}{Numeric threshold of how many mutations to sample.}

\item{source_df_name}{Character name of the data frame object in the environment.}

\item{source_name}{Label for tracking in \code{sampling_summary_df}.}

\item{missing}{Integer number of how many mutations are still missing for the patient.}
}
\value{
Updated \code{missing} 0 how many missing mutations still remain after running the function
}
\description{
Internal helper function that tries to sample additional synonymous mutations from
different sources (patient, global, etc.) until the threshold is reached.
}
